requirements.txt
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.1.5
django-ckeditor==5.9.0
django-cors-headers==3.2.0
django-crispy-forms==1.9.2
django-filter==2.1.0
django-formtools==2.1
django-import-export==2.3.0
django-js-asset==1.2.2
django-redis==4.0.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-extensions==0.6.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
PyJWT==1.7.1
PyMySQL==0.10.0
pytz==2020.1
PyYAML==5.3.1
redis==2.10.6
six==1.15.0
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
